+++ example/server/src/router/load.ts
@@ -1,4 +1,5 @@
 import { get } from 'koa-router-decors';
+import utils from '../utils/index';
 
 export default class Load {
     @get('/load')
@@ -20,6 +21,17 @@ export default class Load {
             result: false,
             errorCode: '',
             errorMsg: '请求失败',
+        };
+    }
+
+    @get('/cloud/rebuildToken')
+    public async rebuildToken(ctx) {
+        ctx.body = {
+            result: true,
+            data: {
+                token: Buffer.from(utils.uuid()).toString('base64'),
+                refreshToken: Buffer.from(utils.uuid()).toString('base64'),
+            },
         };
     }
 }